chore: Update microservice configurations and synchronize routing rules

+++ services/ml-service/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Float, DateTime, BigIntegerField
+from sqlalchemy import Column, Integer, String, Float, DateTime, BigInteger
 from sqlalchemy.ext.declarative import declarative_base
 from datetime import datetime
 
@@ -8,8 +8,8 @@ class FrequentlyBoughtTogether(Base):
     __tablename__ = "frequently_bought_together"
 
     id = Column(Integer, primary_key=True, index=True)
-    product_a_id = Column(BigIntegerField, index=True)
-    product_b_id = Column(BigIntegerField)
+    product_a_id = Column(BigInteger, index=True)
+    product_b_id = Column(BigInteger)
     score = Column(Float)
     computed_at = Column(DateTime, default=datetime.utcnow)
 
@@ -17,16 +17,16 @@ class UserProductView(Base):
     __tablename__ = "user_product_views"
 
     id = Column(Integer, primary_key=True, index=True)
-    user_id = Column(BigIntegerField, index=True)
-    product_id = Column(BigIntegerField, index=True)
+    user_id = Column(BigInteger, index=True)
+    product_id = Column(BigInteger, index=True)
     viewed_at = Column(DateTime, default=datetime.utcnow)
 
 class UserProductInteraction(Base):
     __tablename__ = "user_product_interactions"
 
     id = Column(Integer, primary_key=True, index=True)
-    user_id = Column(BigIntegerField, index=True)
-    product_id = Column(BigIntegerField, index=True)
+    user_id = Column(BigInteger, index=True)
+    product_id = Column(BigInteger, index=True)
     interaction_type = Column(String(50)) # 'view', 'cart', 'purchase'
     timestamp = Column(DateTime, default=datetime.utcnow)
 
@@ -39,3 +39,4 @@ class AnalyticsAggregate(Base):
     value = Column(Float)
     period = Column(String(50)) # 'daily', 'weekly', 'monthly'
     computed_at = Column(DateTime, default=datetime.utcnow)
+
